# binary search definitions for massive speedup