On our massive rails app at AngelList we have 12k ruby files. Running rubrowser took a looong time. I looked into what it was doing when running, and I realized most of the time was spent in == function due to a O(n^2) part of the algorithm that was finding matches of namespaces. Ruby 2 introduced bsearch on enumerable which lets us do the same work with O(n*log n). The only catch is the definitions set needs to be sorted.

The rspec test had to change because it was assuming a certain order of the definitions.

Before patch
771.29s user 0.63s system 100% cpu 12:51.45 total

After patch
60.03s user 0.57s system 99% cpu 1:00.61 total