@emad-elsaid 2021-01-31
Hey, thanks for your improvement have read it now and it looks straight forward. I'll test it later and merge it and release a new version.